loops_packages

+++ Shell-script/16-delete-old-logs.sh
@@ -5,11 +5,17 @@ SOURCE_DIR="/home/ec2-user/shell_script_logs"
 FILES_TO_DELETE=$(find $SOURCE_DIR -name "*.log" -mtime +30)
 echo "$FILES_TO_DELETE"
 
-while read -r files
-do
-    echo "print the files to delete :: $files"
+if [ -f $FILES_TO_DELETE ]
+then
+    echo " files present to delete +30days"
+    while read -r files
+    do
+        echo "print the files to delete :: $files"
 
-   rm -rf $files
+    rm -rf $files
 
-   echo "Files deleted :: $files "
-done <<< $FILES_TO_DELETE
+    echo "Files deleted :: $files "
+    done <<< $FILES_TO_DELETE
+else
+    echo "No files present to delete +30days"
+fi
